+++ src/app/log/log-publishers.service.ts
@@ -1,0 +1,25 @@
+// Source: https://www.codemag.com/article/1711021/Logging-in-Angular-Applications
+
+import {Injectable} from "@angular/core";
+
+import {LogPublisher, LogConsole, LogLocalStorage} from "./log-publishers";
+
+@Injectable()
+export class LogPublishersService {
+  constructor() {
+    // Build publishers arrays
+    this.buildPublishers();
+  }
+
+  // Public properties
+  publishers: LogPublisher[] = [];
+
+  // Build publishers array
+  buildPublishers(): void {
+    // Create instance of LogConsole Class
+    this.publishers.push(new LogConsole());
+
+    // Create instance of LogLocalStorage Class
+    this.publishers.push(new LogLocalStorage());
+  }
+}